+++ consensus_decentralization/metrics/gini.py
@@ -1,7 +1,7 @@
 import logging
 import numpy as np
 import sys
-import src.helper as hlp
+import consensus_decentralization.helper as hlp
 
 
 def compute_gini(blocks_per_entity):
