+++ consensus_decentralization/metrics/herfindahl_hirschman_index.py
@@ -1,6 +1,6 @@
 import logging
 import sys
-import src.helper as hlp
+import consensus_decentralization.helper as hlp
 
 
 def compute_hhi(blocks_per_entity):
