+++ consensus_decentralization/parse.py
@@ -1,9 +1,9 @@
 import argparse
 import logging
-from src.parsers.default_parser import DefaultParser
-from src.parsers.dummy_parser import DummyParser
-from src.parsers.ethereum_parser import EthereumParser
-from src.helper import RAW_DATA_DIR, OUTPUT_DIR
+from consensus_decentralization.parsers.default_parser import DefaultParser
+from consensus_decentralization.parsers.dummy_parser import DummyParser
+from consensus_decentralization.parsers.ethereum_parser import EthereumParser
+from consensus_decentralization.helper import RAW_DATA_DIR, OUTPUT_DIR
 
 
 ledger_parser = {
